Share hook output block parsing
